refactor: extract shared library
