+++ .gitignore
@@ -2,5 +2,3 @@ node_modules
 .idea
 # Ignore built ts files
 __tests__/runner/*
-lib
-dist
