+++ hack/run-e2e-tests.sh
@@ -6,7 +6,6 @@ set -euo pipefail
 [[ -z ${GITHUB_TOKEN} ]] && echo "GITHUB_TOKEN env variable is required. obtail one from https://github.com/settings/tokens/new" && exit 1
 
 yarn build
-yarn package
 
 act pull_request -e tests/act-pull-request.json \
 	-s GITHUB_TOKEN=${GITHUB_TOKEN} \
