+++ src/approver/main.ts
@@ -1,6 +1,6 @@
 import * as core from '@actions/core'
 import * as github from '@actions/github'
-import {Octokit} from './octokit'
+import {Octokit} from '../lib'
 
 async function run(): Promise<void> {
   try {
