--- src/status/utils.ts
@@ -1,7 +1,0 @@
-import type {Status, Job} from './types'
-
-export const isFinalStatus = (st: Status): Boolean =>
-  st === 'failure' || st === 'success'
-
-export const isContributionJob = (job: Job): Boolean =>
-  job.contribution !== undefined
